Added a second surfaceview to render subtittles
The option to pass a second surface view will render all subtitles to
that window, if none passed with merge onto video frame. By default the
window will render to display size adapted to the aspect ratio of the
video and max dimensions only for image subtitles. The subtitle render
frame size can be set later and subtitles will invalidate and rerender
when that occurs.