Added the ability to seek
Added the UI with a seekbar and moved tapping to pause/play to a button
in the center of the screen. The seek function can get the duration
and playback time. The VPlayerListener can get the playback state (pause
and play) and the playback progress to update the interface.

Scrubbing and pause seek are not implemented yet as they require frame
step to be properly implemented.

Also fixed gaining the common metadata from the codec.