Added an api to get bitmaps of a subtitle frame at a certain time
Use ASSRenderer to create a new ASSTrack, then add subtitles to the
track. Finally use the track to render each section of text onto a
ASSBitmap of either ARGB or RGBA and get the data back for rendering
either to a Bitmap or other. This uses libass in native code to render
the subtitle and then use asm to blend it to a buffer which is copied
back to Java to be used.